Use a separate formatter instead of monkey patching RSpec
The monkey patch will wreak havoc when using preloaders like Spring.

To get the specialized formatter, specify it as:

    rspec --format RSpec::ExampleSteps::Formatter
